Progress checkpoint: mouseover event works for a single element
This is due to using querySelector instead of querySelectorAll, the
former of which returns only the first match. Now I need to find
out how I can take the NodeList generated by querySelectorAll,
transform it into an array using Array.from(), and then select
the individual elements of that array corresponding to the event.